Added TraversableOnce
